+++ apps/e2e/action-bars-storybook-e2e/src/e2e/summary-action-bar.component.cy.ts
@@ -138,5 +138,5 @@ describe('action-bars-storybook - summary action bar', () => {
         });
       });
     });
-  }, true);
+  });
 });