+++ apps/e2e/forms-storybook-e2e/src/e2e/field-group.component.cy.ts
@@ -16,5 +16,5 @@ describe('field-group', () => {
         });
       });
     });
-  }, true);
+  });
 });