+++ apps/e2e/indicators-storybook-e2e/src/e2e/wait.component.cy.ts
@@ -41,5 +41,5 @@ describe('indicators-storybook', () => {
         });
       });
     });
-  }, true);
+  });
 });