+++ apps/e2e/layout-storybook-e2e/src/e2e/fluid-grid.component.cy.ts
@@ -26,5 +26,5 @@ describe('layout-storybook - fluid-grid', () => {
         });
       });
     });
-  }, true);
+  });
 });