+++ apps/e2e/pages-storybook-e2e/src/e2e/page-layout-tabs.component.cy.ts
@@ -21,5 +21,5 @@ describe(`pages-storybook`, () => {
         });
       });
     });
-  }, true);
+  });
 });