+++ apps/e2e/theme-storybook-e2e/src/e2e/theming.component.cy.ts
@@ -35,5 +35,5 @@ describe('theme-storybook', () => {
         );
       });
     });
-  }, true);
+  });
 });