+++ apps/playground/src/app/home/home.component.ts
@@ -154,7 +154,7 @@ export class HomeComponent implements AfterViewInit {
             }
           }),
         );
-      } else if (route.data) {
+      } else if (route.data && !this.componentData.includes(route.data)) {
         route.data.path = parentPath + '/' + route.path;
         this.componentData.push(route.data);
         promises.push(Promise.resolve());