+++ libs/components/icon/src/lib/modules/icon/icon-svg-resolver.service.ts
@@ -4,7 +4,7 @@ import { SkyIconVariantType } from './types/icon-variant-type';
 
 async function getIconMap(): Promise<Map<string, number[]>> {
   const response = await fetch(
-    `https://sky.blackbaudcdn.net/static/skyux-icons/7/assets/svg/skyux-icons.svg`,
+    `https://sky.blackbaudcdn.net/static/skyux-icons/8/assets/svg/skyux-icons.svg`,
   );
 
   /* istanbul ignore next */