+++ libs/sdk/eslint-config-skyux/src/index.ts
@@ -158,7 +158,10 @@ const config = tseslint.config(
       '@typescript-eslint/dot-notation': 'error',
       '@typescript-eslint/non-nullable-type-assertion-style': 'error',
       '@typescript-eslint/prefer-includes': 'error',
-      '@typescript-eslint/prefer-nullish-coalescing': 'error',
+      '@typescript-eslint/prefer-nullish-coalescing': [
+        'error',
+        { ignorePrimitives: { bigint: true, number: true, string: true } },
+      ],
       '@typescript-eslint/prefer-optional-chain': 'error',
       '@typescript-eslint/prefer-string-starts-ends-with': 'error',
     },