+++ libs/sdk/skyux-eslint/src/rules/template/prefer-label-text.ts
@@ -1,8 +1,12 @@
-import { TmplAstElement } from '@angular-eslint/bundled-angular-compiler';
+import {
+  TmplAstElement,
+  type TmplAstTextAttribute,
+} from '@angular-eslint/bundled-angular-compiler';
 import {
   ensureTemplateParser,
   getTemplateParserServices,
 } from '@angular-eslint/utils';
+import type { RuleFix, RuleFixer } from '@typescript-eslint/utils/ts-eslint';
 
 import { getChildNodeOf, getTextContent } from '../utils/ast-utils';
 import { createESLintTemplateRule } from '../utils/create-eslint-template-rule';
@@ -56,6 +60,70 @@ const SELECTORS_WITH_LABEL_COMPONENTS = COMPONENTS_WITH_LABEL_TEXT.map(
   (c) => c.selector,
 ).join('|');
 
+const FORM_CONTROL_CLASS = 'sky-form-control';
+
+function getAttributeByName(
+  el: TmplAstElement,
+  attrName: string,
+): TmplAstTextAttribute | undefined {
+  for (const attribute of el.attributes) {
+    if (attribute.name === attrName) {
+      return attribute;
+    }
+  }
+
+  return;
+}
+
+/**
+ * Removes the "sky-form-control" CSS class from <sky-input-box /> input elements
+ * to satisfy the input box's input directive selector.
+ * See: https://github.com/blackbaud/skyux/blob/040e461a50cb3d08ff8f7332dba350b7e97c5fd8/libs/components/forms/src/lib/modules/input-box/input-box-control.directive.ts#L11
+ */
+function removeFormControlClass(
+  fixer: RuleFixer,
+  inputEl: TmplAstElement,
+): RuleFix[] {
+  const fixes: RuleFix[] = [];
+
+  const classAttr = getAttributeByName(inputEl, 'class');
+
+  if (classAttr?.value.includes(FORM_CONTROL_CLASS)) {
+    const classnames = classAttr.value;
+
+    // Remove "class" attribute if it only contains the form control class.
+    if (classnames === FORM_CONTROL_CLASS) {
+      fixes.push(
+        fixer.removeRange([
+          classAttr.sourceSpan.start.offset,
+          classAttr.sourceSpan.end.offset,
+        ]),
+      );
+    } else {
+      const index = classnames.indexOf(FORM_CONTROL_CLASS);
+      const classLength = FORM_CONTROL_CLASS.length;
+
+      /* istanbul ignore else: safety check */
+      if (classAttr.valueSpan) {
+        let classStart = classAttr.valueSpan.start.offset + index;
+        let classEnd = classStart + classLength;
+
+        // Account for multiple classes in the "class" attribute, and remove extra
+        // whitespace.
+        if (classnames.at(index - 1) === ' ') {
+          classStart--;
+        } else if (classnames.at(index + classLength) === ' ') {
+          classEnd++;
+        }
+
+        fixes.push(fixer.removeRange([classStart, classEnd]));
+      }
+    }
+  }
+
+  return fixes;
+}
+
 export const rule = createESLintTemplateRule({
   create(context) {
     ensureTemplateParser(context);
@@ -81,9 +149,25 @@ export const rule = createESLintTemplateRule({
           el.inputs.some((i) => i.name === labelInputName) ||
           el.attributes.some((i) => i.name === labelInputName);
 
-        const labelEl = getChildNodeOf(el, labelSelector);
+        const labelEl = getChildNodeOf(el, [labelSelector]);
+        const inputEl = getChildNodeOf(el, ['input', 'select', 'textarea']);
 
         if (labelEl) {
+          // Abort if the `skyId` directive is used on a child input of the
+          // `<sky-input-box />` component since its inclusion usually means the
+          // user wishes to implement "hard mode".
+          if (
+            el.name === 'sky-input-box' &&
+            inputEl &&
+            getAttributeByName(inputEl, 'skyId')
+          ) {
+            return;
+          }
+
+          const hasElementChildren = labelEl.children.some(
+            (child) => child instanceof TmplAstElement,
+          );
+
           context.report({
             loc: parserServices.convertNodeSourceSpanToLoc(el.sourceSpan),
             messageId,
@@ -92,29 +176,36 @@ export const rule = createESLintTemplateRule({
               labelInputName,
               labelSelector,
             },
-            fix: (fixer) => {
-              const textContent = getTextContent(labelEl);
-              const textReplacement = ` ${labelInputName}="${textContent}"`;
-              const range = [
-                el.startSourceSpan.end.offset - 1,
-                el.startSourceSpan.end.offset,
-              ] as const;
+            // Don't fix if the label includes child elements.
+            fix: hasElementChildren
+              ? undefined
+              : (fixer): RuleFix[] => {
+                  const textContent = getTextContent(labelEl);
+                  const textReplacement = ` ${labelInputName}="${textContent}"`;
+                  const range = [
+                    el.startSourceSpan.end.offset - 1,
+                    el.startSourceSpan.end.offset,
+                  ] as const;
 
-              const fixers = [
-                fixer.removeRange([
-                  labelEl.sourceSpan.start.offset,
-                  labelEl.sourceSpan.end.offset,
-                ]),
-              ];
+                  const fixers = [
+                    fixer.removeRange([
+                      labelEl.sourceSpan.start.offset,
+                      labelEl.sourceSpan.end.offset,
+                    ]),
+                  ];
 
-              if (!hasLabelText) {
-                fixers.push(
-                  fixer.insertTextBeforeRange(range, textReplacement),
-                );
-              }
+                  if (!hasLabelText) {
+                    fixers.push(
+                      fixer.insertTextBeforeRange(range, textReplacement),
+                    );
+                  }
 
-              return fixers;
-            },
+                  if (el.name === 'sky-input-box' && inputEl) {
+                    fixers.push(...removeFormControlClass(fixer, inputEl));
+                  }
+
+                  return fixers;
+                },
           });
         }
       },